requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
Pillow
jinja2
python-multipart